+++ go2cj-new/trainset/programs/21_for_index_value.go
@@ -5,7 +5,6 @@ import "fmt"
 func main() {
 	xs := []int{10, 20, 30}
 	for i, v := range xs {
-		fmt.Println(i)
-		fmt.Println(v)
+		fmt.Println(i, v)
 	}
 }